Fix failing line2info function (SA tag read processing)
Changed line split from split('\n') to split(). The former didn't allow index accessing in the subsequent line.
Changed .append to .extend to add gene info to the gene dictionary. The former resulted in a list of lists that couldn't be processed when creating the gene alignment dictionary.
Adding genesdict as a parameter to line2info. Without this line2info complained that it didn't exist.